Домашнее занятие. Я не очень понял последнее задание так как в таблице Orders не указано какие предметы внутри, только финальная цена. Поэтому ее я и решил проверить, что если мы сложим цены полученные из таблице products по Product id полученными в таблице OrderItems, мы проверим что корректные предметы были положены в order и финальная цена корректная

+++ TestAQA1/DTO/DapperTestsDTO/ProductDTO.cs
@@ -13,7 +13,7 @@ namespace Tests1.DTO.DapperTestsDTO
 
         string description,
 
-        long price,
+        double   price,
 
         long stock,
